+++ accounts/apps.py
@@ -1,10 +1,10 @@
 from django.apps import AppConfig
 
 
-class AccountConfig(AppConfig):
+class AccountsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'account'
+    name = 'accounts'
     verbose_name = "حساب کاربری"
 
     def ready(self):
-        import account.signals
+        import accounts.signals
